backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.6
py==1.11.0
pycparser==2.22
PyJWT==2.9.0
pytest==7.0.0
pytest-django==4.10.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
